pyproject.toml
[project]
name = "game-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Luiz Ricardo", email = "luizrgfelipe@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pillow>=10.0.0",
    "mss>=9.0.0",
    "python-dotenv>=1.0.0",
    "edge-tts>=6.1.0",
]

[project.scripts]
game-agent = "game_agent:main"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
